wip: move route props to declaration
