+++ src/routing/declaration/route-props.type.ts
